+++ processor/src/test/resources/expected/com/example/callback/CustomReturnTypeBuilder.java
@@ -1,0 +1,51 @@
+package com.example.callback;
+
+import java.util.Objects;
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import jsinterop.base.Js;
+import jsinterop.base.JsPropertyMap;
+import react4j.core.React;
+import react4j.core.ReactNode;
+
+@Generated("react4j.processor.ReactProcessor")
+class CustomReturnTypeBuilder {
+  private CustomReturnTypeBuilder() {
+  }
+
+  @Nonnull
+  static ReactNode key(@Nonnull final String key) {
+    return new Builder().key( key );
+  }
+
+  @Nonnull
+  static ReactNode build() {
+    return new Builder().build();
+  }
+
+  public interface Builder1 {
+    @Nonnull
+    ReactNode key(@Nonnull String key);
+  }
+
+  public interface Builder2 {
+    @Nonnull
+    ReactNode build();
+  }
+
+  private static class Builder implements Builder1, Builder2 {
+    private final JsPropertyMap<Object> _props = JsPropertyMap.of();
+
+    @Override
+    @Nonnull
+    public final ReactNode key(@Nonnull final String key) {
+      _props.set( "key", Objects.requireNonNull( key ) );
+      return build();
+    }
+
+    @Nonnull
+    public final ReactNode build() {
+      return React.createElement( React4j_CustomReturnType.TYPE, Js.uncheckedCast( _props ) );
+    }
+  }
+}