+++ processor/src/test/resources/expected/com/example/callback/TypeParameterOnCallback_.java
@@ -2,7 +2,6 @@ package com.example.callback;
 
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
-import react4j.annotations.Callback;
 
 @Generated("react4j.processor.ReactProcessor")
 class TypeParameterOnCallback_ {
@@ -10,7 +9,7 @@ class TypeParameterOnCallback_ {
   }
 
   @Nonnull
-  static Callback.Procedure _handleFoo(@Nonnull final TypeParameterOnCallback component) {
+  static TypeParameterOnCallback.CustomHandler _handleFoo(@Nonnull final TypeParameterOnCallback component) {
     return ((React4j_TypeParameterOnCallback) component)._handleFoo;
   }
 }